Create https environment option
Now, in the default environment, you can specify

```
ssl: true,
key: '/path/to/key',
cert: '/path/to/cert'
```

in order to enable HTTP in Wompt.  In anticipation for https being enabled,
this also removes the strict protocol for scripts in a few of the templates
and layout partials (changing from ```http://...``` to ```//...```, as it is
supported in a fair amount of browsers these days).